Initial commit of basic BST functionality, with unit tests.
More testing could be done to verify, but it looks like the basic functionality
for inserting and searching the tree works. There's a lot of code duplication,
as searching and inserting are 90% the same, but are here implemented completely
seperately.

The rank attribute exists on the node struct, but isn't used as of yet. This is
for use in the balanced tree later.