Added Clear and Destroy Extended the BST API to include functions for clearing (free all nodes) the bst and for freeing the bst. Additionally, a combination clear_destroy function was added that does both.
Clearing is done via a post-order tree traversal. It frees the node objects,
but doesn't do anything beyond that.

The test code was updated to call clear_destroy at the end to free the test
tree, but no further tests were added to verify that it was fully and properly
freed.