Create a "tracker" module
Moved the rank update tracking routines out to their own module,
along with a few other general cleanup items.

Additionally, added balance_factor to bstnode. This is needed for
AVL, and in the interest of code reuse, I'm going to set the AVL
tree up using the bstnode and bst datatypes. It should be able to
reuse most of the bst_ functions, aside from insert and delete.