BST Rotations
I implemented left and right rotations about a specified center
node, which will be used to maintain balance during insertions
and deletions. For testing, the avl code currently just uses
standard bst insert and delete--the next step is of course to
replace those with routines that track the balance and use the
rotations to correct it.