Insert/Delete Path Tracking
Updated the tracker for rank updates to be a more generic tracker
of the path followed for insertions/deletions. This allows the use
of the same generic "find location and path" routine for both
inserts and deletes, facilitates rolling back rank updates as needed,
and will also allow easy use of an auxillary stack for balancing the
avl tree. Knuth reccomends using such a stack for delete, so that was
the impetus for this. I may update insert to use the stack as well,
for consistency between the two routines, although it is less
efficient than what I have now.