Updated Insert to use Stack
Updated the insert routine to use a stack to track the update path,
and make the necessary balance updates/rotations based on this. This
is technically a little less efficient than the original way, but it
is a lot cleaner and also reflects more closely the procedure that
delete will use.