Moved Rotations to bst and Updated delete
Moved the rotation operations from AVL to BST, and then updated
the way that BST delete works.

bst_delete now uses left_rotate to move the node to be deleted down
the tree until it has no right children, and then deletes it using the
normal approach. This is a lot less efficient, but the code is far
simpler now, and I'm hoping that I can use this for avl_delete by just
taking steps to rebalance the tree after each left_rotate.