Balanced Delete Setup
Initial setup of my idea for making avl_node_delete ensure that the
tree below the element to be deleted is balanced, in spite of the
rotations used to get the deleted node lower in the tree.

Once the tree below the original location of the deleted node is
balanced, the current avl_delete balancing should be able to handle
balancing the tree above that point.